inject the newly created goosh.js.compr into the goosh.html file

+++ deploy.sh
@@ -13,12 +13,22 @@ echo "gzip"
 
 cat goosh.js.compr.tmp |sed s\#gshell.grothkopp.com\#goosh.org\#g |sed s\#ABQIAAAA0cXSEVCNSwf_x74KTtPJMRShYK5vgJfK0afUKMRqjECszDItkhTOIyZ74499O_ys5nJIQuP4sq4nZg\#ABQIAAAA0cXSEVCNSwf_x74KTtPJMRQP4Q7D8MPck7bhT7upyfJTzVDU2BRxkUdd2AvzlDDF7DNUJI_Y4eB6Ug\#g >goosh.js.compr
 
-wget 'http://n.goosh.org/?deploy=1' -O goosh.html -o /dev/null
+wget 'http://n.goosh.org/?deploy=1' -O goosh.html-dl -o /dev/null
+
+scriptLine=`grep -n "var goosh=new Object();" goosh.html-dl | cut -d : -f 1`
+
+cat goosh.html-dl | head -n $(($scriptLine-1)) > goosh.html
+cat goosh.js.compr >> goosh.html
+cat goosh.html-dl | tail -n +$(($scriptLine+1)) >> goosh.html
 
 #exit
 echo "copy files"
 
-cp ../goosh.org/index.html ../goosh.org/index.html-autosave
+if [ ! -f ../goosh.org/index.html ]; then
+  mkdir ../goosh.org
+else
+  cp ../goosh.org/index.html ../goosh.org/index.html-autosave
+fi
 cp goosh.html ../goosh.org/index.html
 #echo "gzip"
